refactor: move Axi4CrossbarTester2
